requirements.txt
docopt==0.6.2
pyserial==3.5